refactor(stores): decompose usePilotStore into 4 slices (max-lines fix)
Drops usePilotStore.ts from 470 effective LOC (over the 400 lint
budget) to 121 raw / ~80 effective. The store body is now a thin
composition layer that delegates to:

  - usePilotStore.types.ts     (151 LOC) — shared interface contract
  - usePilotStore.api.ts       (212 LOC) — REST/CRUD: load/create/update/delete
  - usePilotStore.skills.ts    (234 LOC) — skill / wound / ability mutations
  - usePilotStore.selectors.ts ( 51 LOC) — useFilteredPilots / usePilotById

Public hook surface is unchanged: usePilotStore, useFilteredPilots,
usePilotById remain importable from '@/stores/usePilotStore'. All 40
existing pilot-store tests pass without modification — behavior is
byte-for-byte identical, only the call sites moved.

Bonus: extracted two private helpers (postCreatePilot,
postWithSuccess) to DRY the near-identical try/catch ladders that
five of the original inline actions shared.
